nao-medical-translator/backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
groq
python-dotenv
pydantic
requests